refactor(cli): move onboarding preflight and inference setup
